Create SpaceCountFunction
This is a simple function to count the number of white spaces in between words in a given cell. The function takes one argument which is the cell containing the string. This function is helpful when you are parsing out and separating first names, middle names, last names, and suffixes. You can use the results of this function on a list of names in conjunction with the Sort feature to sort a list of names by the number of white spaces they contain. Then you can use the Text to Columns feature to work on chunks of names with the same number of spaces.